Refactor test.py and test_workflow package to use common packages
Move git_repository from build_workflow into a new "git" package and update build.py and test.py to use the new location.

Remove the duplicate GitRepository from test_workflow.

Change test.py to use the manifests and git packages. Move bundle_manifest to the manifests package.

Rename integ_test.py to integ_test_suite.py.
